Introduce utils, for methods shared between client and sentinel (originally 274f8462 by Tanguy Le Barzic <tanguy@Zorro2.(none)>, without sentinel.js)
