+++ index.js
@@ -2,6 +2,7 @@
 
 var net = require("net"),
     util = require("./lib/util"),
+    reply_to_object = require("./lib/utils").reply_to_object,
     Queue = require("./lib/queue"),
     to_array = require("./lib/to_array"),
     events = require("events"),
@@ -535,23 +536,6 @@ function try_callback(callback, reply) {
             throw err;
         });
     }
-}
-
-// hgetall converts its replies to an Object.  If the reply is empty, null is returned.
-function reply_to_object(reply) {
-    var obj = {}, j, jl, key, val;
-
-    if (reply.length === 0) {
-        return null;
-    }
-
-    for (j = 0, jl = reply.length; j < jl; j += 2) {
-        key = reply[j].toString();
-        val = reply[j + 1];
-        obj[key] = val;
-    }
-
-    return obj;
 }
 
 function reply_to_strings(reply) {
